8.2.22. рефакторит компонент City.
