+++ project/src/types/city.ts
@@ -1,9 +1,9 @@
 import {Location} from './offer';
 
-export type City = {
+export type CityType = {
   name: string;
   location: Location;
   id: number;
 };
 
-export type Cities = City[];
+export type CitiesType = CityType[];
